pyproject.toml
[project]
name = "mynion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "strands-agents>=1.15.0",
    "uvicorn[standard]>=0.38.0",
    # Slack integration dependencies
    "slack-bolt>=1.21.4",
    "boto3>=1.37.8",
    "aws-cdk-lib>=2.233.0",
    "constructs>=10.4.3",
    "aws-cdk-aws-bedrock-agentcore-alpha>=2.233.0a0",
    # MCP Gateway integration
    "mcp-proxy-for-aws>=0.1.0",
    # AgentCore Runtime SDK
    "bedrock-agentcore>=0.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "ruff>=0.14.5",
    "types-boto3>=1.40.74",
]

[tool.mypy]
python_version = "3.11"
# CDKはcdk/ディレクトリ内で実行されるため、モジュール解決パスを追加
mypy_path = ["cdk"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    "cdk\\.out",
    "cdk/cdk\\.out",
    "\\.venv",
]

[[tool.mypy.overrides]]
module = [
    "strands.*",
    "aws_cdk.*",
    "mcp_proxy_for_aws.*",
    "bedrock_agentcore.*",
    "google.*",
    "googleapiclient.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "cdk.out",
    "cdk/cdk.out",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
